[mstreg] Moving reglib classes into a module
As part of mstflint integration, the reglib related classes moved to
a new folder to be used as a lib by other tools

Issue: 3226599
